refactor(gameStatus): split the summary, and collapse what it wrote twice
GameStatusSummary.tsx was 734 lines and 14 components. It goes to 284.
gameStatusText.ts takes the pure string helpers, Scoreline.tsx takes the
readout with `Center` renamed to `Scoreline`, and useScorelineFit.ts
takes the hook. The hook stays beside the component rather than moving
to src/hooks/, because it queries `game-status__*` class names, which is
component knowledge.

Three collapses ride along. `outcomeClasses` replaces the identical
conditional map in `Score` and `Side`, which the SCSS had already
factored into an `outcome` mixin. `Marker`'s two branches become one
span. `GameMark`'s four branches become one shape plus a `markFor`
lookup, in the order the four `if` statements tested in.

observeResize.ts replaces the block copied between the scoreline fit and
the filler rows. Its `typeof === "undefined"` guard had never run under
any test, because setupTests.ts stubs ResizeObserver, so
observeResize.test.ts deletes and restores the global to exercise it.

Proved by built HTML rather than by reasoning. A throwaway harness
rendered 8 summary states and 5 mark states before and after. The only
diff is attribute order on the blank marker, which is unavoidable when
two JSX branches become one. With attributes sorted the two dumps are
byte-identical.

The unlit marker case was written and seen passing before the collapse.
Nothing asserted that room-holding behavior until now.

573 tests, up from 557. No existing test needed editing.
